Update step_logger
Test release of a logger step_logger
